Add project access token back to api
